Added structured testing with graphable results
Updated the ExampleApp to take a size of list to test, sample size (number of tests to run per 'take'), and an 'increment' that is used to increase the number of items to take per test sample.  Executed tests on list sizes of 100, 1,000, 10,000, and 100,000.  Saved results to be graphed later.